Only prioritize critical input threads
Input has several threads:
InputReader
InputClassifier (InputProcessor)
InputDispatcher
InputFilter
TraceProcessor

However, only two of them are in the critical path of the event
dispatch, and therefore, should be prioritized.

The others do not need to run with high priority.

In this CL, require that all InputThread's declare themselves as
critical or not, which would allow us to set their priority
appropriately.

Bug: 330719044
Test: perfetto trace
Flag: EXEMPT bugfix
Change-Id: Iaa7d7ed32557e1617884590562c3474b17d3a7f9