Reject inconsistent globally injected events
The dispatcher currently crashes when certain inconsistent input events
are injected. This is affecting test stability negatively.

In this CL, we reject globally-injected inconsistent events. That may cause
some test flakiness, but should eliminate the crashes due to dispatcher
reaching bad state later.

Unfortunately, we can't currently reject all inconsistent injected
events. In the case of targeted injection, it is common for the caller
to leave pointers dangling. Since the injection happens into the
caller-owned windows only, at the end of those tests the windows get
cleaned up, so the dispatcher is still in a good state.

The eventual goal is to completely get rid of injection. Meanwhile,
however, this should help avoid at least some of the crashes.

Bug: 369935405
Flag: EXEMPT bugfix
Test: atest inputflinger_tests
Change-Id: I0696dbd3e4c5b88aad5aa853759227c0b56d5374