SF: Remove EventLog frame duration
In the process of replacing FrameTracker it became apparent that there
was a good deal of supporting logic for frame duration. This logic was
flawed (all high numbers go to the last bucket) and is not ideal to
replicate in the new FrameTimeline based future.

By removing this, FrameTracker replacement is significantly easier.

Bug: 241394120
Flag: EXEMPT EventLog removal
Test: presubmit
Change-Id: I526561b11c21cd2872e9f128849b33883ae9c948