SF: Trace input info whenever it is present, even if input token is null
The bug is that if inputInfo.token is null, the function will return early without writing anything to the proto. This is incorrect, as the proto should still be written even if inputInfo.token is null, since the input
window can have a touchable region that can affect input dispatching.

For example, the various "input sink" windows usually have input info,
but set the NO_INPUT_CHANNEL flag and thus have null input channels
and tokens.

This CL fixes the bug by removing the if statement that checks if inputInfo.token is null.

Bug: 373401882
Change-Id: Ide2ebf4a932139dac766101bb52921ff64b02899
Test: manual: take input trace with winscope, and look for input sinks
Flag: EXEMPT bug fix