Skip primaryRangeIsSingleRate check for ARR
If the active mode is ARR enabled, then this means the physical
refresh rate is a single rate that does not change but the device can
render at any of the divisors of the vsync rate.
Therefore doesn't make sense to perform
the primaryRangeIsSingleRate check for ARR mode.

Flag: EXEMPT bugfix
Test: atest RefreshRateSelectorTest
Test: atest FrameRateOverrideTest
Bug: 308020695
Bug: 369287577

Change-Id: Ie16b1561aef05287cdac1908bfe9fd6fd1ab43fa