Implement NDK createSessionUsingConfig API
This patch implements a new way to create ADPF hint session,
with a new parcelable SessionCreationConfig. This object encapsulates
the required information for session creation, with a set of setters to
assign values to those fields in the session creation config.

This object can be reused on the client side in order to streamline
the session creation process.

Bug: 362801903
Bug: 367803904
Test: atest HintManagerServiceTest
      atest PerformanceHintNativeTest
Flag: EXEMPT_NDK
Change-Id: Idd2ea87886c664b625f40066cc05c6df1bf0e6d4