EventHub: log error when getting axis info for non-existent device
It's possible that getAbsoluteAxisInfo not finding the device is the
cause of a hard-to-reproduce crash in GestureConverter, where nullopt is
returned for an axis despite the EventHub classification code requiring
that axis to classify a device as a touchpad. Trying to retrieve axis
info for a non-existent device is always an error of some kind, so let's
at least log that.

Bug: 374255515
Test: m checkinput
Flag: EXEMPT log only update
Change-Id: Ifb0c82d88c11ac31671e947d5f5805949931060f