Consume full tap in EndToEndNativeInputTest
Currently, the injection inside the test is happening asynchronously.
Occasionally, it seems that an inconsistent event stream is being sent.
Without further debugging about how this is possible, a speculative fix
is being made here by consuming a complete motion sequence after a tap
is injected.

This way, we know that the test must wait for the UP event to be
received, thus hopefully preventing the next test from running and
starting the asynchronous injection.

To prevent misuse of 'consumeEvent' in the future, it is also being made
private in this CL. Unfortunately, a bunch of other fields should also
be private (but were not made so in the beginning, possibly a typo). As
a result, some of the fields are now relied upon. These are kept as is
to keep focus of this CL.

Bug: 361264974
Test: atest libgui_test
Flag: TEST_ONLY
Change-Id: I99bac4022bab2b3aa106e0838b227f24eb512ea1