Remove superfluous #include.
This is implicit in Android's sys/cdefs.h and this header is also
built for the host, which doesn't have this header. The missing header
was previously hacked around by symlinking the header into a directory
that would be included by host builds, but the real header is
sensitive to the contents of sys/cdefs.h so that wasn't sound anyway
(and we in bionic have no interest in keeping it and its dependencies
compatible with non-Android).

It was included for __INTRODUCED_IN(), but that's already defined by a
command-line argument, so we can just remove the #include.

Bug: None
Test: treehugger
Change-Id: Ia9b057118ac70de1a8d2d6b8898b9fee00887a7e